+++ Client/src/main/java/com/entity/Client.java
@@ -2,10 +2,15 @@ package com.entity;
 
 import org.springframework.stereotype.Component;
 
-//@Component
+@Component
 public class Client {
     private final String clientName;
     private boolean ignored;
+
+    public Client() {
+        this.clientName = "";
+        ignored = false;
+    }
 
     public Client(String clientName) {
         this.clientName = clientName;
