+++ Client/src/main/java/com/entity/User.java
@@ -1,10 +1,11 @@
 package com.entity;
 
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
+import org.springframework.stereotype.Component;
 
+import java.util.ArrayList;
+import java.util.List;
+
+@Component
 public class User {
     private String userName;
     private String password;
