+++ Client/src/main/resources/application.properties
@@ -5,3 +5,4 @@ ServerName=localhost
 ServerPort=1436
 ProxyHost=51.38.28.127
 ProxyPort=80
+spring.main.allow-circular-references=true
