+++ Client/src/main/resources/com/controller/bottom.fxml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.*?>
+<BorderPane xmlns="http://javafx.com/javafx"
+            xmlns:fx="http://javafx.com/fxml"
+            fx:controller="com.controller.BottomController"
+            BorderPane.alignment="CENTER">
+    <left>
+        <Label prefHeight="19.0" prefWidth="118.0" styleClass="chat-label"
+               stylesheets="@../../stylesheets/chatclient.css" text="General Message!"
+               BorderPane.alignment="CENTER"/>
+    </left>
+    <right>
+        <HBox BorderPane.alignment="CENTER">
+            <Button fx:id="btnSend" disable="true" mnemonicParsing="false" onAction="#btnHandler"
+                    styleClass="chat-button" stylesheets="@../../stylesheets/chatclient.css"
+                    text="Send Message!"/>
+            <Button mnemonicParsing="false" onAction="#btnHandler" styleClass="chat-button"
+                    stylesheets="@../../stylesheets/chatclient.css" text="Exit Chat"/>
+        </HBox>
+    </right>
+    <center>
+        <TextField fx:id="txtMessage" onAction="#txtHandler" prefHeight="26.0" prefWidth="483.0"
+                   promptText="Broadcast Message" BorderPane.alignment="CENTER"/>
+    </center>
+</BorderPane>
