+++ Client/src/main/resources/com/controller/center.fxml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.ScrollPane?>
+<?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.text.TextFlow?>
+<ScrollPane xmlns:fx="http://javafx.com/fxml/1"
+            xmlns="http://javafx.com/javafx/18"
+            fx:id="sp_main"
+            fx:controller="com.controller.CenterController"
+            prefHeight="200.0" prefWidth="200.0" style="-fx-background-color: white;"
+            BorderPane.alignment="CENTER">
+    <TextFlow fx:id="messageBoard" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308"
+              prefHeight="433.0" prefWidth="574.0" style="-fx-background-color: white;"/>
+</ScrollPane>
