+++ Client/src/main/resources/com/controller/tab/userstab.fxml
@@ -5,8 +5,8 @@
 <?import javafx.scene.control.ScrollPane?>
 <?import javafx.scene.layout.BorderPane?>
 <?import javafx.scene.layout.VBox?>
-<BorderPane xmlns="http://javafx.com/javafx/17.0.2-ea"
-            xmlns:fx="http://javafx.com/fxml/1"
+<BorderPane xmlns:fx="http://javafx.com/fxml/1"
+            xmlns="http://javafx.com/javafx/17.0.2-ea"
             fx:controller="com.controller.tab.UsersTabController"
             style="-fx-font-weight: bold; -fx-font-color: white;">
     <center>
