+++ Server/src/main/resources/com/controller/server.fxml
@@ -60,7 +60,7 @@
                     <Insets bottom="10.0" left="10.0" right="10.0" top="10.0"/>
                 </GridPane.margin>
             </TextField>
-            <TextField fx:id="txtMaximumGuest"  GridPane.columnIndex="1" GridPane.rowIndex="2">
+            <TextField fx:id="txtMaximumGuest" GridPane.columnIndex="1" GridPane.rowIndex="2">
                 <GridPane.margin>
                     <Insets bottom="10.0" left="10.0" right="10.0" top="10.0"/>
                 </GridPane.margin>
