+++ server.properties
@@ -1,5 +1,5 @@
 #Amir Chat v1.0 Amir Technologies (USA) Pvt Ltd..
-#Fri Feb 16 08:54:46 EST 2024
+#Sat Feb 17 12:44:10 EST 2024
 MaximumGuest=50
 PortNumber=1436
 ServerName=1436
